refactor(cluster): split WorkerRemove into worker_remove.go
- move WorkerRemove and removeNodesConf from worker.go to
  worker_remove.go
- worker.go retains the add flow, shared helpers (findController,
  writeNodesConfAndReconfigure), and cleanup logic
- tests stay in worker_test.go due to shared test infrastructure
