+++ pkg/cluster/worker.go
@@ -169,77 +169,6 @@ func WorkerAdd(ctx context.Context, client *docker.Client, meshMgr *mesh.Manager
 	return nodes, nil
 }
 
-// WorkerRemove removes worker nodes from a cluster.
-//
-// For managed nodes (those present in sind-nodes.conf), the flow is:
-//  1. Update sind-nodes.conf to remove the node definitions
-//  2. Reconfigure slurmctld
-//  3. Deregister DNS + known_hosts
-//  4. Stop + remove containers
-//
-// For unmanaged nodes, only steps 3–4 are performed.
-func WorkerRemove(ctx context.Context, client *docker.Client, meshMgr *mesh.Manager, clusterName string, shortNames []string) error {
-	log := sindlog.From(ctx)
-	realm := meshMgr.Realm
-
-	if len(shortNames) == 0 {
-		return nil
-	}
-
-	log.InfoContext(ctx, "removing workers", "cluster", clusterName, "nodes", strings.Join(shortNames, ","))
-
-	// List cluster containers to find controller and validate targets.
-	containers, err := client.ListContainers(ctx, "label="+LabelCluster+"="+clusterName)
-	if err != nil {
-		return fmt.Errorf("listing containers: %w", err)
-	}
-
-	controller, hasController := findController(containers, realm, clusterName)
-	containerMap := make(map[docker.ContainerName]docker.ContainerListEntry, len(containers))
-	for _, c := range containers {
-		containerMap[c.Name] = c
-	}
-
-	// Resolve which nodes to remove, checking they exist and are worker nodes.
-	seen := make(map[string]bool, len(shortNames))
-	var targets []docker.ContainerListEntry
-	for _, name := range shortNames {
-		if seen[name] {
-			continue
-		}
-		seen[name] = true
-		cn := ContainerName(realm, clusterName, name)
-		c, ok := containerMap[cn]
-		if !ok {
-			return fmt.Errorf("node %q not found in cluster %q", name, clusterName)
-		}
-		if role := c.Labels[LabelRole]; role != "worker" {
-			return fmt.Errorf("node %q has role %q: only worker nodes can be removed with worker remove", name, role)
-		}
-		targets = append(targets, c)
-	}
-
-	// For managed nodes: check if sind-nodes.conf exists and update it.
-	if hasController {
-		nodesConf, err := client.ReadFile(ctx, controller.Name, "/etc/slurm/sind-nodes.conf")
-		if err == nil {
-			// sind-nodes.conf exists → remove managed nodes from it.
-			if err := removeNodesConf(ctx, client, controller.Name, nodesConf, shortNames); err != nil {
-				return err
-			}
-		}
-		// If ReadFile fails, sind-nodes.conf doesn't exist → treat as unmanaged.
-	}
-
-	// Deregister DNS + known_hosts.
-	if err := DeregisterMesh(ctx, meshMgr, clusterName, targets); err != nil {
-		return err
-	}
-
-	// Stop + remove containers.
-	return DeleteContainers(ctx, client, targets)
-}
-
 // ValidateWorkerAdd checks prerequisites for adding workers to a cluster.
 // For managed workers, it verifies that sind-nodes.conf exists on the
 // controller (indicating sind-generated Slurm configuration is in use).
@@ -343,13 +272,6 @@ func updateNodesConf(ctx context.Context, client *docker.Client, controllerName 
 	}
 	updated := slurm.AddNodesToConf(current, entries)
 
-	return writeNodesConfAndReconfigure(ctx, client, controllerName, updated)
-}
-
-// removeNodesConf removes node definitions from sind-nodes.conf and
-// reconfigures slurmctld.
-func removeNodesConf(ctx context.Context, client *docker.Client, controllerName docker.ContainerName, currentConf string, shortNames []string) error {
-	updated := slurm.RemoveNodesFromConf(currentConf, shortNames)
 	return writeNodesConfAndReconfigure(ctx, client, controllerName, updated)
 }
